requirements.txt
torch
transformers
sklearn
numpy
pillow
jsonlines
pyyaml
